requirements.txt
fastapi>=0.100.0 # Using a recent version
uvicorn[standard]>=0.20.0
bleak>=0.20.0
Pillow>=9.0.0
jinja2>=3.0.0
python-multipart>=0.0.5 # For form data/file uploads